feat(src): extract youtube_embed module and refactor existing code
- Added new module `youtube_embed` for handling YouTube embed operations.
- Refactored `YoutubeState` to include methods for retrieving Invidious client and base URL.
- Removed unused dependencies on `quick_xml`, `regex`, and related regex patterns from `src/youtube.rs`.
- Updated route handlers to utilize the new methods in `YoutubeState` for fetching embed configurations and proxying requests.
- Simplified HTML URL rewriting logic by removing static regex patterns and replacing with more dynamic approaches.
- Removed functionality related to DASH manifest rewriting and quality indicator script injection, as these are now handled by the new module.
